Ciclo-Circular-main/requirements-prod.txt
django
psycopg2-binary
dj-database-url
python-dotenv
djangorestframework
django-admin-interface
django-colorfield
django-crispy-forms
django-tables2
django-filter
django-extensions
django-import-export
django-anymail
whitenoise
django-allauth
PyJWT
cryptography
openpyxl
xhtml2pdf==0.2.11
python-Levenshtein
rapidfuzz
pillow
requests
gunicorn
google-generativeai
openai
pymupdf
python-docx
mercadopago
qrcode
matplotlib
pandas
numpy
scipy
scikit-learn
python-bidi==0.4.2
wordcloud